requirements.txt
python-decouple
beautifulsoup4
selenium
PIL
numpy
pandas